# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI_Chatbot API Skeleton
  version: 0.1.0-skeleton
  description: Auto-generated skeleton from workbook; schemas are intentionally coarse.
servers:
- url: http://localhost:8080
paths:
...
  /v1/ops/llm/providers/{provider}/kill-switch:
    post:
      operationId: opsProviderKillswitch
      summary: LLM Provider Kill-Switch
      description: Immediate action API to stop routing to a specific model provider.
      tags:
      - 운영 관측 통합
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: OPS-PROVIDER-KILLSWITCH
      x-source-row: 73
...
  /v1/ops/rollbacks:
    post:
      operationId: opsRollbackTrigger
      summary: Policy/Prompt/KB Rollback Trigger
      description: Immediate action API to trigger rollback for policy/prompt/KB versions.
      tags:
      - 운영 관측 통합
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: OPS-ROLLBACK-TRIGGER
      x-source-row: 74
  /v1/ops/traces:
    get:
      operationId: opsTraceQuery
      summary: 추적 로그 조회
      description: 추적 로그 조회 기능 제공
      tags:
      - 운영 관측 통합
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: OPS-TRACE-QUERY
      x-source-row: 53
  /v1/ops/workflow/reports:
    get:
      operationId: intWorkflowReport
      summary: 통합 워크플로 리포트
      description: 통합 워크플로 리포트 기능 제공
      tags:
      - 운영 관측 통합
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
      - name: to
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Workbook skeleton response (200)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - OPS
      x-program-id: INT-WORKFLOW-REPORT
      x-source-row: 61
  /v1/rag/answer:
    post:
      operationId: apiRagAnswer
      summary: 근거 기반 답변 생성
      description: 근거 기반 답변 생성 기능 제공
      tags:
      - RAG 검색 응답
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: API-RAG-ANSWER
      x-source-row: 20
...
  /v1/rag/clarify/suggest:
    post:
      operationId: apiRagClarifySuggest
      summary: 명확화 질문 제안
      description: 명확화 질문 제안 기능 제공
      tags:
      - RAG 검색 응답
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: API-RAG-CLARIFY-SUGGEST
      x-source-row: 23
  /v1/rag/query/classify:
    post:
      operationId: apiRagQueryClassify
      summary: 질의 의도 분류
      description: 질의 의도 분류 기능 제공
      tags:
      - RAG 검색 응답
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: API-RAG-QUERY-CLASSIFY
      x-source-row: 22
  /v1/rag/retrieve:
    post:
      operationId: apiRagRetrieve
      summary: RAG 검색
      description: RAG 검색 기능 제공
      tags:
      - RAG 검색 응답
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: API-RAG-RETRIEVE
      x-source-row: 19
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT